Implementing Delete Owner

+++ MyVet.Web/Controllers/PetTypesController.cs
@@ -19,13 +19,11 @@ namespace MyVet.Web.Controllers
             _context = context;
         }
 
-        // GET: PetTypes
         public async Task<IActionResult> Index()
         {
             return View(await _context.PetTypes.ToListAsync());
         }
 
-        // GET: PetTypes/Details/5
         public async Task<IActionResult> Details(int? id)
         {
             if (id == null)
@@ -43,15 +41,11 @@ namespace MyVet.Web.Controllers
             return View(petType);
         }
 
-        // GET: PetTypes/Create
         public IActionResult Create()
         {
             return View();
         }
 
-        // POST: PetTypes/Create
-        // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
-        // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("Id,Name")] PetType petType)
@@ -65,7 +59,6 @@ namespace MyVet.Web.Controllers
             return View(petType);
         }
 
-        // GET: PetTypes/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
             if (id == null)
@@ -81,9 +74,6 @@ namespace MyVet.Web.Controllers
             return View(petType);
         }
 
-        // POST: PetTypes/Edit/5
-        // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
-        // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, [Bind("Id,Name")] PetType petType)
@@ -116,7 +106,6 @@ namespace MyVet.Web.Controllers
             return View(petType);
         }
 
-        // GET: PetTypes/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
             if (id == null)
@@ -125,21 +114,19 @@ namespace MyVet.Web.Controllers
             }
 
             var petType = await _context.PetTypes
-                .FirstOrDefaultAsync(m => m.Id == id);
+                .Include(pt => pt.Pets)
+                .FirstOrDefaultAsync(pt => pt.Id == id);
             if (petType == null)
             {
                 return NotFound();
             }
 
-            return View(petType);
-        }
+            if (petType.Pets.Count > 0)
+            {
+                ModelState.AddModelError(string.Empty, "The pet type can't be removed.");
+                return RedirectToAction(nameof(Index));
+            }
 
-        // POST: PetTypes/Delete/5
-        [HttpPost, ActionName("Delete")]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> DeleteConfirmed(int id)
-        {
-            var petType = await _context.PetTypes.FindAsync(id);
             _context.PetTypes.Remove(petType);
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
